Added Frustum Compute Data Structure, refactored code to have GPU Manager
