+++ data/config.json
@@ -1,7 +1,7 @@
 {
     "num_additional_threads_specified": 14,
     "use_all_system_threads": true,
-    "render_radius": 5,
+    "render_radius": 0,
     "atlas_max_images_on_a_row": 16,
     "atlas_max_images_on_a_column": 16,
     "max_amount_of_blocktypes": 65536,
