+++ hello-wgpu/src/gpu_manager/flag_state.rs
@@ -1,0 +1,4 @@
+pub struct FlagState {
+    pub should_calculate_frustum: bool,
+    pub render_wireframe: bool,
+}
