+++ hello-wgpu/src/gpu_manager/gpu_data/vertex_gpu_data.rs
@@ -1,11 +1,31 @@
 use std::collections::HashMap;
-
+use cgmath::Point3;
+use derivables::vertex::Vertex;
 use fundamentals::world_position::WorldPosition;
-use wgpu::{Device, util::DeviceExt};
+use wgpu::{Device, util::DeviceExt, BufferUsages};
 
-use crate::{voxels::mesh::Mesh, camera::Camera};
+use crate::voxels::mesh::Mesh;
 
 use super::vec_vertex_index_length_triple::VecVertexIndexLengthsTriple;
+
+pub const NUM_BUCKETS: usize = (fundamentals::consts::NUMBER_OF_CHUNKS_AROUND_PLAYER as usize) * fundamentals::consts::NUM_BUCKETS_PER_CHUNK;
+
+pub struct BucketData {
+    pub vertex_buffer_number: usize,
+    pub vertex_offset: usize,
+    pub index_buffer_number: usize,
+    pub index_offset: usize,
+    pub num_indices: usize
+}
+
+pub struct MeshBucketData {
+    pub front_bucket_data_indices: Vec<u32>,
+    pub back_bucket_data_indices: Vec<u32>,
+    pub left_bucket_data_indices: Vec<u32>,
+    pub right_bucket_data_indices: Vec<u32>,
+    pub top_bucket_data_indices: Vec<u32>,
+    pub bottom_bucket_data_indices: Vec<u32>,
+}
 
 pub struct VertexGPUData {
     pub data_front: VecVertexIndexLengthsTriple,
@@ -14,18 +34,24 @@ pub struct VertexGPUData {
     pub data_right: VecVertexIndexLengthsTriple,
     pub data_top: VecVertexIndexLengthsTriple,
     pub data_bottom: VecVertexIndexLengthsTriple,
-    pub chunk_index_array: [WorldPosition; fundamentals::consts::NUMBER_OF_CHUNKS_AROUND_PLAYER as usize],
+    pub chunk_index_array: Vec<WorldPosition>,
     pub chunk_index_buffer: wgpu::Buffer,
     pub chunk_index_bind_group_layout: wgpu::BindGroupLayout,
     pub chunk_index_bind_group: wgpu::BindGroup,
     pub pos_to_gpu_index: HashMap<WorldPosition, usize>,
     pub pos_to_loaded_index: HashMap<WorldPosition, usize>,
-    pub loaded_chunks: Vec<WorldPosition>
+    pub loaded_chunks: Vec<WorldPosition>,
+    //Vertex Pooling
+    pub vertex_pool_buffers: Vec<wgpu::Buffer>,
+    pub index_pool_buffers: Vec<wgpu::Buffer>,
+    pub indirect_pool_buffers: Vec<wgpu::Buffer>,
+    pub vertex_count_pool_buffers: Vec<wgpu::Buffer>,
+    pub pool_position_to_mesh_bucket_data: HashMap<WorldPosition, MeshBucketData>,
 }
 
 impl VertexGPUData {
-    pub fn new(camera: &Camera, device: &Device) -> Self {
-        let chunk_index_array = fundamentals::consts::get_positions_around_player(WorldPosition::from(camera.position));
+    pub fn new(camera_position: Point3<f32>, device: &Device) -> Self {
+        let chunk_index_array = fundamentals::consts::get_positions_around_player(WorldPosition::from(camera_position));
         let mut pos_to_gpu_index = HashMap::new();
         for (i, chunk_position) in chunk_index_array.iter().enumerate() {
             pos_to_gpu_index.insert(*chunk_position, i);
@@ -62,6 +88,69 @@ impl VertexGPUData {
             ],
             label: Some("chunk_index_bind_group")
         });
+
+        const MAX_BUFFER_SIZE: usize = 1073741824;
+
+        let vertex_bucket_size = std::mem::size_of::<Vertex>() * fundamentals::consts::NUM_VERTICES_IN_BUCKET as usize;
+        let number_of_vertex_buckets_per_buffer = MAX_BUFFER_SIZE / vertex_bucket_size;
+        let mut number_of_vertex_pool_buffers = NUM_BUCKETS / number_of_vertex_buckets_per_buffer;
+        if NUM_BUCKETS % number_of_vertex_buckets_per_buffer != 0 {
+            number_of_vertex_pool_buffers += 1;
+        }
+
+        // * 3/2 = 6/4 because for every 4 vertices there are 6 indices
+        let index_bucket_size = std::mem::size_of::<i32>() * fundamentals::consts::NUM_VERTICES_IN_BUCKET as usize * 3 / 2;
+        let number_of_index_buckets_per_buffer = MAX_BUFFER_SIZE / index_bucket_size;
+        let mut number_of_index_pool_buffers = NUM_BUCKETS / number_of_index_buckets_per_buffer;
+        if NUM_BUCKETS % number_of_index_buckets_per_buffer != 0 {
+            number_of_index_pool_buffers += 1;
+        }
+
+        let number_of_buckets_per_buffer = std::cmp::min(number_of_vertex_buckets_per_buffer, number_of_index_buckets_per_buffer);
+        let number_of_buffers = std::cmp::max(number_of_vertex_pool_buffers, number_of_index_pool_buffers);
+
+        let mut vertex_pool_buffers = Vec::new();
+        let mut index_pool_buffers = Vec::new();
+        let mut indirect_pool_buffers = Vec::new();
+        let mut vertex_count_pool_buffers = Vec::new();
+
+        for i in 0..number_of_buffers {
+            let pool_vertex_buffer = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some(format!("Vertex Pool Buffer {i}").as_str()),
+                size: (number_of_buckets_per_buffer * vertex_bucket_size) as u64,
+                usage: BufferUsages::VERTEX | BufferUsages::COPY_DST | BufferUsages::STORAGE,
+                mapped_at_creation: false
+            });
+
+            let pool_index_buffer = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some(format!("Index Pool Buffer {i}").as_str()),
+                size: (number_of_buckets_per_buffer * index_bucket_size) as u64,
+                usage: BufferUsages::INDEX | BufferUsages::COPY_DST | BufferUsages::STORAGE,
+                mapped_at_creation: false
+            });
+
+            let pool_indirect_buffer = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some(format!("Indirect Pool Buffer {i}").as_str()),
+                size: (number_of_buckets_per_buffer * (std::mem::size_of::<u32>() * 5)) as u64,
+                usage: BufferUsages::INDIRECT | BufferUsages::COPY_DST | BufferUsages::STORAGE,
+                mapped_at_creation: false
+            });
+
+            let pool_vertex_count_buffer = device.create_buffer(&wgpu::BufferDescriptor {
+                label: Some(format!("Vertex Count Pool Buffer {i}").as_str()),
+                size: (number_of_buckets_per_buffer * (std::mem::size_of::<u32>())) as u64,
+                usage: BufferUsages::COPY_DST | BufferUsages::STORAGE,
+                mapped_at_creation: false
+            });
+
+            vertex_pool_buffers.push(pool_vertex_buffer);
+            index_pool_buffers.push(pool_index_buffer);
+            indirect_pool_buffers.push(pool_indirect_buffer);
+            vertex_count_pool_buffers.push(pool_vertex_count_buffer);
+        }
+
+        let pool_position_to_mesh_bucket_data = HashMap::new();
+
         Self {
             data_front: VecVertexIndexLengthsTriple::new(),
             data_back: VecVertexIndexLengthsTriple::new(),
@@ -75,7 +164,12 @@ impl VertexGPUData {
             chunk_index_buffer,
             chunk_index_bind_group_layout,
             chunk_index_bind_group,
-            loaded_chunks: Vec::new()
+            loaded_chunks: Vec::new(),
+            vertex_pool_buffers,
+            index_pool_buffers,
+            indirect_pool_buffers,
+            vertex_count_pool_buffers,
+            pool_position_to_mesh_bucket_data,
         }
     }
