+++ hello-wgpu/src/state/flag_state.rs
@@ -1,5 +1,3 @@
 pub struct FlagState {
     pub mouse_pressed: bool,
-    pub calculate_frustum: bool,
-    pub render_wireframe: bool,
 }
