--- hello-wgpu/src/state/render_state.rs
@@ -1,4 +1,0 @@
-pub struct RenderState {
-    pub render_pipeline_regular: wgpu::RenderPipeline,
-    pub render_pipeline_wireframe: wgpu::RenderPipeline,
-}
