--- hello-wgpu/src/state/texture_state.rs
@@ -1,7 +1,0 @@
-use crate::texture;
-
-pub struct TextureState {
-    pub diffuse_bind_group: wgpu::BindGroup,
-    pub diffuse_texture: texture::Texture,
-    pub depth_texture: texture::Texture,
-}
